this pr moves a bunch of files around, removes the mainpage page and centralizes everything in the index files
