+++ client/data/sections.ts
@@ -1,0 +1,9 @@
+const sections = [
+  { title: "About Me", id: 1 },
+  { title: "Experience", id: 2 },
+  { title: "Open Source", id: 3 },
+  { title: "Technologies", id: 4 },
+  // { title: "Projects", id: 5 },
+  // { title: "Contact", id: 6 },
+];
+export default sections;
