+++ client/data/technologiesInfo.ts
