+++ client/pages/_app.tsx
@@ -1,5 +1,6 @@
 import "@/styles/globals.css";
 import type { AppProps } from "next/app";
+import "../styles/customStyles.css";
 
 export default function App({ Component, pageProps }: AppProps) {
   return <Component {...pageProps} />;
